+++ todo/src/main/java/bonux/yada/todo/services/YadaServices.java
@@ -24,38 +24,3 @@ public interface YadaServices {
     void delete(Integer id);
 }
 
-@Service
-class YadaServicesImpl
-        implements YadaServices {
-    @Autowired
-    private TodoRepo todoRepo;
-
-    @Override
-    public void delete(Integer id) {
-        todoRepo.deleteById(id);
-    }
-
-    @Override
-    public Optional<Todo> update(Integer id,
-                                 Todo updateTodo) {
-        var maybeTodo = todoRepo
-                .findById(id)
-                .map(todo -> ModelTodoBuilder.from(todo).update(updateTodo).build());
-        return maybeTodo.map(todo -> DomainTodoBuilder.fromModel(todoRepo.update(todo)));
-    }
-
-    @Override
-    public Todo create(Todo todo) {
-        return DomainTodoBuilder.fromModel(todoRepo.create(ModelTodoBuilder.builder().create(todo).build()));
-    }
-
-    @Override
-    public Optional<Todo> get(Integer id) {
-        return todoRepo.findById(id).map(DomainTodoBuilder::fromModel);
-    }
-
-    @Override
-    public Stream<Todo> get() {
-        return todoRepo.findAll().stream().map(DomainTodoBuilder::fromModel);
-    }
-}
